fix: MedGemma float16 NaN crash on MPS — use float32 for inference
float16 causes "probability tensor contains inf/nan" during softmax
sampling on Apple Silicon MPS. Switched to float32 which is stable.
Also added try/except to generate endpoint for proper error reporting.

+++ apps/backend/api/services/medgemma.py
@@ -75,9 +75,10 @@ class MedGemmaService:
 
             logger.info(f"Using device: {self.device}")
 
-            # float16 is faster than bfloat16 on MPS (native Metal fp16 support)
+            # float32 on MPS — float16 causes NaN in softmax during sampling
+            # (known PyTorch MPS numerical instability with half-precision LLMs)
             if self.device == "mps":
-                dtype = torch.float16
+                dtype = torch.float32
             elif self.device == "cuda":
                 dtype = torch.bfloat16
             else:
